refactor(Alert, Confirm, Prompt Modal): 폴더 구조 변경
