+++ components/src/lib/AlertModal/AlertModal.stories.tsx
@@ -1,11 +1,11 @@
 import '../../../index.css';
 import { useState } from 'react';
-import { ConfirmModal, Modal } from '../../';
+import { AlertModal, Modal } from '..';
 import type { Meta, StoryObj } from '@storybook/react';
 
-const meta: Meta<typeof ConfirmModal> = {
-  title: 'Modal/ConfirmModal',
-  component: ConfirmModal,
+const meta: Meta<typeof AlertModal> = {
+  title: 'Modal/AlertModal',
+  component: AlertModal,
   decorators: [
     (Story) => (
       <div style={{ width: '100vw', height: '100vh' }}>
@@ -20,22 +20,23 @@ const meta: Meta<typeof ConfirmModal> = {
     const [isOpen, setIsOpen] = useState(true);
 
     const onClose = () => setIsOpen(false);
-    const onConfirm = () => console.log('confirm');
+    const onConfirm = () => {
+      console.log('confirm');
+    };
 
     return (
       <>
         <Modal.Button onClick={() => setIsOpen(true)} size='small' backgroundColor='primary'>
           모달열기
         </Modal.Button>
-        <ConfirmModal {...args} isOpen={isOpen} onClose={onClose} onConfirm={onConfirm} />
+        <AlertModal {...args} isOpen={isOpen} onClose={onClose} onConfirm={onConfirm} />
       </>
     );
   },
 };
 
 export default meta;
-
-type Story = StoryObj<typeof ConfirmModal>;
+type Story = StoryObj<typeof AlertModal>;
 
 export const Default: Story = {
   args: {
@@ -44,6 +45,7 @@ export const Default: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
+    isOpen: true,
   },
 };
 
@@ -54,6 +56,7 @@ export const Bottom: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
+    isOpen: true,
   },
 };
 
@@ -64,6 +67,7 @@ export const Center: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
+    isOpen: true,
   },
 };
 
@@ -74,6 +78,7 @@ export const Small: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
+    isOpen: true,
   },
 };
 
@@ -84,6 +89,7 @@ export const Medium: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
+    isOpen: true,
   },
 };
 
@@ -94,6 +100,7 @@ export const Large: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
+    isOpen: true,
   },
 };
 
@@ -106,7 +113,6 @@ export const ExistCloseButton: Story = {
     existCloseButton: true,
   },
 };
-
 export const ExcludeCloseButton: Story = {
   args: {
     position: 'center',
