+++ components/src/lib/ConfirmModal/ConfirmModal.stories.tsx
@@ -1,11 +1,11 @@
 import '../../../index.css';
 import { useState } from 'react';
-import { AlertModal, Modal } from '../../';
+import { ConfirmModal, Modal } from '..';
 import type { Meta, StoryObj } from '@storybook/react';
 
-const meta: Meta<typeof AlertModal> = {
-  title: 'Modal/AlertModal',
-  component: AlertModal,
+const meta: Meta<typeof ConfirmModal> = {
+  title: 'Modal/ConfirmModal',
+  component: ConfirmModal,
   decorators: [
     (Story) => (
       <div style={{ width: '100vw', height: '100vh' }}>
@@ -20,23 +20,22 @@ const meta: Meta<typeof AlertModal> = {
     const [isOpen, setIsOpen] = useState(true);
 
     const onClose = () => setIsOpen(false);
-    const onConfirm = () => {
-      console.log('confirm');
-    };
+    const onConfirm = () => console.log('confirm');
 
     return (
       <>
         <Modal.Button onClick={() => setIsOpen(true)} size='small' backgroundColor='primary'>
           모달열기
         </Modal.Button>
-        <AlertModal {...args} isOpen={isOpen} onClose={onClose} onConfirm={onConfirm} />
+        <ConfirmModal {...args} isOpen={isOpen} onClose={onClose} onConfirm={onConfirm} />
       </>
     );
   },
 };
 
 export default meta;
-type Story = StoryObj<typeof AlertModal>;
+
+type Story = StoryObj<typeof ConfirmModal>;
 
 export const Default: Story = {
   args: {
@@ -45,7 +44,6 @@ export const Default: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
-    isOpen: true,
   },
 };
 
@@ -56,7 +54,6 @@ export const Bottom: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
-    isOpen: true,
   },
 };
 
@@ -67,7 +64,6 @@ export const Center: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
-    isOpen: true,
   },
 };
 
@@ -78,7 +74,6 @@ export const Small: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
-    isOpen: true,
   },
 };
 
@@ -89,7 +84,6 @@ export const Medium: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
-    isOpen: true,
   },
 };
 
@@ -100,7 +94,6 @@ export const Large: Story = {
     title: '제목을 입력해주세요',
     label: '설명을 입력해주세요.',
     existCloseButton: true,
-    isOpen: true,
   },
 };
 
@@ -113,6 +106,7 @@ export const ExistCloseButton: Story = {
     existCloseButton: true,
   },
 };
+
 export const ExcludeCloseButton: Story = {
   args: {
     position: 'center',
